Support loading entrypoints by passing a module instead of a function.
The old way of defining salt loader entrypoints was, for example, on `setup.cfg`:
```ini
[options.entry_points]
salt.loader=
  runner_dirs = thirpartypackage.loader:func_to_get_list_of_dirs
  module_dirs = thirpartypackage.loader:func_to_get_list_of_dirs
```
This old way is still supported in the rare ocasion where you actually
want to pass multiple paths to the loader and to not break backwards
compatibility.

The new way of defining the salt loader entrypoints is, for example, on `setup.cfg`:
```ini
[options.entry_points]
salt.loader=
  <this-name-does-not-matter = thirpartypackage
  <this-name-does-not-matter = thirpartypackage:callable
```

With this new approach, when the target is a package, Salt looks for the
`ext_type` being loaded as submodule of the package and if it exists, it
loads from there.
If the target is a callable, then, that callable should return a
dictionary where the keys are the several salt loader names. For
salt execution modules it would be `modules`, for salt state modules it would
be `states`, etc. The same directory layout you see in salt.
The values for those keys needs to be an iterable(not a string which is
also an iterable), ie, a list, a generator, etc, whose values are the
paths from where to load the salt loader modules.

Fixes #58939